+++ src/ppscore/__init__.py
@@ -1,14 +1,13 @@
 # -*- coding: utf-8 -*-
-from pkg_resources import get_distribution, DistributionNotFound
+from importlib.metadata import version, PackageNotFoundError
 
 try:
     # Change here if project is renamed and does not equal the package name
     dist_name = __name__
-    __version__ = get_distribution(dist_name).version
-except DistributionNotFound:  # pragma: no cover
+    __version__ = version(dist_name)
+except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
 finally:
-    del get_distribution, DistributionNotFound
-
+    del version, PackageNotFoundError
 
 from ppscore.calculation import score, predictors, matrix